object_segmentation/requirements.txt
pyserial==3.5
requests==2.31.0
PyYAML==6.0.1
azure-core==1.30.1
opencv-python==4.10.0.82
ultralytics==8.2.84
pymodbus==3.7.0
wandb==0.16.1
imgaug==0.4.0
Pillow==9.3.0
tqdm==4.66.1
pip==21.2.4
colorlog==6.8.0
configparser==6.0.0
wheel==0.41.2
PySocks==1.7.1
protobuf==3.20.0
cryptography==42.0.5
lxml==4.9.3
tornado==6.3.3
beautifulsoup4==4.12.2
bleach==6.0.0
pytz==2023.3
isodate==0.6.1
platformdirs==3.10.0
attrs==24.1.0
httpx==0.24.1
requests-file==2.1.0
certifi==2023.7.22
six==1.16.0
idna==3.4
urllib3==1.26.16
charset-normalizer==3.2.0
simplejson==3.19.1
setuptools==69.5.1
future==0.18.3
pydicom==2.4.4
python-multipart==0.0.9